chore(database, logic-state-indexer): Split database/postgres/state_indexer into files. Remove redundant block_hash from handle_state_changes method in logic-state-indexer
